Added Create To build.gradle and split commands up a lil
