+++ gradle.properties
@@ -18,9 +18,9 @@ minecraft_version=1.21.1
 # as they do not follow standard versioning conventions.
 minecraft_version_range=[1.21.1,1.22)
 # The Neo version must agree with the Minecraft version to get a valid artifact
-neo_version=21.1.141
+neo_version=21.1.222
 # The Neo version range can use any version of Neo as bounds
-neo_version_range=[21.1.141,)
+neo_version_range=[21.1.222,)
 # The loader version range can only use the major version of FML as bounds
 loader_version_range=[1,)
 
@@ -43,3 +43,13 @@ mod_group_id=com.createcivilization.capitol
 mod_authors=McArctic, Mavity The Madity, Orion
 # The description of the mod. This is a simple multiline text string that is used for display purposes in the mod list.
 mod_description=Teams and claims systems for Minecraft.
+
+# JourneyMap
+journeymap_api_version=2.0.0-1.21.1-SNAPSHOT
+journeymap_curse_file_id=6689206
+
+# Create Mod
+create_version=6.0.9-215
+ponder_version=1.0.81
+flywheel_version=1.0.6
+registrate_version=MC1.21-1.3.0+67
