+++ src/main/java/com/createcivilization/capitol/Capitol.java
@@ -5,11 +5,13 @@ import com.createcivilization.capitol.common.managers.DatabaseManager;
 import com.createcivilization.capitol.server.commands.CapitolCommands;
 import com.createcivilization.capitol.common.networking.CapitolNetworking;
 
+import com.createcivilization.capitol.server.events.CreateEvents;
 import com.mojang.logging.LogUtils;
 
 import net.minecraft.world.level.storage.LevelResource;
 import net.neoforged.bus.api.IEventBus;
 import net.neoforged.fml.ModContainer;
+import net.neoforged.fml.ModList;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.fml.loading.FMLEnvironment;
 
@@ -41,6 +43,10 @@ public class Capitol {
 
 		CapitolCommands.init(NeoForge.EVENT_BUS);
 		CapitolNetworking.init(modEventBus);
+
+//		if (ModList.get().isLoaded("create")) {
+//			NeoForge.EVENT_BUS.register(CreateEvents.class);
+//		}
     }
