+++ src/main/java/com/createcivilization/capitol/server/events/CreateEvents.java
@@ -1,0 +1,9 @@
+package com.createcivilization.capitol.server.events;
+
+import net.neoforged.bus.api.SubscribeEvent;
+
+public class CreateEvents {
+
+
+
+}
