+++ src/main/resources/META-INF/neoforge.mods.toml
@@ -23,6 +23,13 @@ versionRange = "${minecraft_version_range}"
 ordering = "NONE"
 side = "BOTH"
 
+[[dependencies."${mod_id}"]]
+modId = "create"
+type = "optional"
+versionRange = "[6.0.9,6.1.0)"
+ordering = "NONE"
+side = "BOTH"
+
 #[[dependencies."${mod_id}"]]
 #modId = "journeymap"
 #mandatory = false
